web/requirements.txt
quart
requests
pyyaml